made the nameID submission optional for the creation of aspect on context; centralized the generation of nameID into the namingService
