+++ src/domain/context/aspect/index.ts
@@ -1,5 +1,5 @@
 export * from './aspect.entity';
 export * from './aspect.interface';
-export * from './dto/aspect.dto.create';
+export * from '../context/dto/context.dto.create.aspect';
 export * from './dto/aspect.dto.update';
 export * from './dto/aspect.dto.delete';
