+++ src/domain/context/context/context.resolver.mutations.ts
@@ -2,7 +2,6 @@ import { Args, Mutation, Resolver } from '@nestjs/graphql';
 import { CurrentUser, Profiling } from '@src/common/decorators';
 import { IReference } from '@domain/common/reference';
 import { ContextService } from '@domain/context/context/context.service';
-import { CreateAspectInput, IAspect } from '@domain/context/aspect';
 import { AuthorizationPrivilege } from '@common/enums';
 import { GraphqlGuard } from '@core/authorization';
 import { UseGuards } from '@nestjs/common/decorators';
@@ -10,17 +9,17 @@ import { AuthorizationService } from '@core/authorization/authorization.service'
 import { AgentInfo } from '@core/authentication';
 import { CreateReferenceOnContextInput } from '@domain/context/context/dto/context.dto.create.reference';
 import { ReferenceService } from '@domain/common/reference/reference.service';
-import { AspectService } from '@domain/context/aspect/aspect.service';
 import { AuthorizationPolicyService } from '@domain/common/authorization-policy/authorization.policy.service';
 import { CreateCanvasOnContextInput } from './dto/context.dto.create.canvas';
 import { ICanvas } from '@domain/common/canvas';
 import { CanvasAuthorizationService } from '@domain/common/canvas/canvas.service.authorization';
 import { DeleteCanvasOnContextInput } from './dto/context.dto.delete.canvas';
 import { AspectAuthorizationService } from '../aspect/aspect.service.authorization';
+import { CreateAspectOnContextInput } from './dto/context.dto.create.aspect';
+import { IAspect } from '../aspect/aspect.interface';
 @Resolver()
 export class ContextResolverMutations {
   constructor(
-    private aspectService: AspectService,
     private referenceService: ReferenceService,
     private authorizationPolicyService: AuthorizationPolicyService,
     private authorizationService: AuthorizationService,
@@ -63,7 +62,7 @@ export class ContextResolverMutations {
   @Profiling.api
   async createAspectOnContext(
     @CurrentUser() agentInfo: AgentInfo,
-    @Args('aspectData') aspectData: CreateAspectInput
+    @Args('aspectData') aspectData: CreateAspectOnContextInput
   ): Promise<IAspect> {
     const context = await this.contextService.getContextOrFail(
       aspectData.contextID
