+++ src/domain/context/context/dto/context.dto.create.aspect.ts
@@ -1,0 +1,9 @@
+import { UUID } from '@domain/common/scalars';
+import { CreateAspectInput } from '@domain/context/aspect/dto/aspect.dto.create';
+import { InputType, Field } from '@nestjs/graphql';
+
+@InputType()
+export class CreateAspectOnContextInput extends CreateAspectInput {
+  @Field(() => UUID, { nullable: false })
+  contextID!: string;
+}
